+++ index.html
@@ -15,6 +15,6 @@
 		<div class="allNotes"></div>
 	</body>
 
-	<script src="./js/main.js"></script>
-	<!-- <script src="./js/utils/addNote.js"></script> -->
+	<!-- <script type="module" src="./src/generateID.js"></script> -->
+	<script type="module" src="./js/main.js"></script>
 </html>
